# frontend/openapi/openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "PLIEGO API",
    "description": "API de PLIEGO v1: autenticación, clientes, catálogo, inventario, carrito, pago y pedidos.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18445",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/verify-email": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Verificar correo",
        "description": "Consume un token de uso único que vence a las 24 horas. No inicia sesión automáticamente.",
        "operationId": "verify",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Correo verificado"
          },
          "400": {
            "description": "Enlace inválido, vencido o utilizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Restablecer contraseña",
        "description": "Consume token de recuperación, aplica BCrypt y revoca todas las sesiones de refresco. Los access tokens anteriores conservan su expiración máxima de 30 minutos.",
        "operationId": "reset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Contraseña restablecida"
          },
          "400": {
            "description": "Enlace o contraseña inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/resend-verification": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Solicitar otro enlace de verificación",
        "description": "Respuesta neutra. Un minuto entre emisiones, hasta cinco solicitudes por correo y veinte por IP por hora.",
        "operationId": "resend",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Solicitud recibida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/forgot-password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Solicitar recuperación de contraseña",
        "description": "Respuesta idéntica para cuentas existentes, desconocidas, bloqueadas o limitadas. Enlace de uso único durante 15 minutos.",
        "operationId": "forgot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Solicitud recibida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetail": {
        "type": "object",
        "description": "Detalle de problema RFC 9457 con campos de correlación de PLIEGO.",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "urn:pliego:problem:VALIDATION_ERROR"
          },
          "title": {
            "type": "string",
            "example": "Datos inválidos"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Revisa los datos enviados e intenta nuevamente."
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "example": "/api/v1/auth/register"
          },
          "code": {
            "type": "string",
            "description": "Código estable legible por máquina. Los errores de dominio usan el SQLSTATE canónico de PostgreSQL (Pxxxx); los errores propios de la capa REST usan un código simbólico (por ejemplo VALIDATION_ERROR o INVALID_CARD_NUMBER).",
            "example": "P3002"
          },
          "traceId": {
            "type": "string",
            "example": "8f2ab598-3e33-40b1-a5c7-15f676a84f9a"
          }
        }
      },
...
      "EmailVerificationRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "pattern": "[A-Za-z0-9_-]{43}"
          }
        },
        "required": [
          "token"
        ]
      },
      "PasswordResetRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "pattern": "[A-Za-z0-9_-]{43}"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        },
        "required": [
          "password",
          "token"
        ]
      },
      "EmailActionRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254,
            "minLength": 0
          }
        },
        "required": [
          "email"
        ]
      },
      "AcceptedResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "authSessionCookie": {
        "type": "apiKey",
        "name": "pliego_session",
        "in": "cookie"
      }
    }
  }
}